requirements.txt
# Core ML Libraries
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# API Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0

# Azure ML
azureml-core==1.51.0
azureml-sdk==1.51.0
azure-storage-blob==12.17.0
azure-identity==1.14.0

# MLflow for model tracking
mlflow==2.6.0
mlflow-skinny==2.6.0

# Data validation and drift detection
evidently==0.4.5
great-expectations==0.17.15

# Monitoring
prometheus-client==0.17.1
psutil==5.9.5

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0

# Development and Testing
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Jupyter for exploration
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2

# Optional: DVC for data versioning
dvc==3.15.0
dvc-azure==2.22.0